+++ packages/meteor-components/plugins/mixins/plugin.js
@@ -23,20 +23,13 @@ Component.onComponentInitialize(function (component, template) {
     });
 
     template.onCreated(function () {
-      let tplInst = this;
-
       mixins.forEach(function (mixin) {
-        mixin.templateInstance = tplInst;
         if (mixin.onCreated) mixin.onCreated();
       });
     });
 
     template.onRendered(function () {
-      let tplInst = this;
-      let refs = Component.getRefs(tplInst);
-
       mixins.forEach(function (mixin) {
-        mixin.refs = Object.create(refs);
         if (mixin.onRendered) mixin.onRendered();
       });
     });
@@ -47,8 +40,6 @@ Component.onComponentInitialize(function (component, template) {
         let mixin = mixinInstances.pop();
         if (mixin.onDestroyed) mixin.onDestroyed();
         mixin.owner = null;
-        mixin.refs = null;
-        mixin.templateInstance = null;
       }
     });
   }